release: Fix release process
Previously the release process relied heavily on the pipeline.
This moves some of the work locally, which makes the proccess a
bit more manual but less magic.

The old release processes also required the e2e test to run as
well as pushed Docker images. This required secrets in the
repository, which are no longer available.

Testing of the releases need to be done separately until we have
introduced automation.